Linux/requirements.txt
customtkinter
Pillow
getmac
asyncio
ipaddress
python-nmap
paramiko
pymetasploit3
scapy